requirements.txt
asgiref==3.2.7
astroid==2.3.3
colorama==0.4.3
Django==3.0.5
django-classy-tags==1.0.0
django-comments-xtd==2.4.2
django-contrib-comments==1.9.2
django-crispy-forms==1.9.0
django-disqus==0.5
django-taggit==1.2.0
django-taggit-templatetags2==1.6.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
docutils==0.16
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mock==4.0.2
Pillow==7.1.1
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
wrapt==1.11.2
